style: use single quoted string and f-string

+++ src/backend/apps/api/modules/user.py
@@ -12,14 +12,13 @@ class User(object):
             if user.pswd_hash != pswd_hash:
                 raise WrongPasswordException('Wrong password.')
         except ObjectDoesNotExist:
-            raise UserDoNotExistException(
-                'User (Email=%s) do not exist.' % email)
+            raise UserDoesNotExistException(
+                    'User (Email=%s) do not exist.' % email)
 
-        db_user.UserSession.objects.filter(user_id=user).delete()
+        db_user.UserSession.objects.filter(user_id=user.user_id).delete()
 
         self.user_dbobj = user
-        self.user_session_dbobj = db_user.UserSession.objects.create(
-            user_id=user)
+        self.user_session_dbobj = db_user.UserSession.objects.create(user_id=user)
 
         # some code ...
         # self.user_info = UserInfo(...)
@@ -33,23 +32,24 @@ class User(object):
     @classmethod
     def new_user(cls, email, pswd_hash, user_name, gender, resident_city_id):
         if db_user.User.objects.filter(email=email).exists():
-            raise UserAlreadyExistsException(
-                'User already exists, try to login.')
+            raise UserAlreadyExistsException('User already exists, try to login.')
         resident_city = db_city.City.objects.get(city_id=resident_city_id)
-        user = db_user.User.objects.create(email=email, pswd_hash=pswd_hash)
-        user_info = db_user.UserInfo.objects.create(
-            user_id=user, user_name=user_name, gender=gender, resident_city_id=resident_city)
+        user = db_user.User.objects.create(email=email,
+                                           pswd_hash=pswd_hash)
+        user_info = db_user.UserInfo.objects.create(user_id=user,
+                                                    user_name=user_name,
+                                                    gender=gender,
+                                                    resident_city_id=resident_city)
         return cls(email=email, pswd_hash=pswd_hash)
 
 
 class UserInfoBase(object):
     def __init__(self, user_id):
         try:
-            userinfo = db_user.UserInfo.objects.get(user_id=user_id)
+            user_info = db_user.UserInfo.objects.get(user_id=user_id)
         except ObjectDoesNotExist:
-            raise UserDoNotExistException(
-                'User (ID=%s) do not exist.' % user_id)
-        self.user_info_dbobj = userinfo
+            raise UserDoesNotExistException(f'User (ID={user_id}) does not exist.')
+        self.user_info_dbobj = user_info
 
     def get_user_id(self):
         return self.user_info_dbobj.user_id
@@ -60,13 +60,12 @@ class UserInfoBase(object):
     def get_gender(self):
         return self.user_info_dbobj.gender
 
-    def get_resident_city(self):
+    def get_resident_city_id(self):
         try:
-            city = db_city.City.objects.get(
-                city_id=self.user_info_dbobj.resident_city_id)
+            city = db_city.City.objects.get(city_id=self.user_info_dbobj.resident_city_id)
         except ObjectDoesNotExist:
-            raise CityIdDoNotExistException(
-                'City whose ID=%s do not exist.' % self.user_info_dbobj.resident_city_id)
+            raise CityIdDoesNotExistException(f'City (ID={self.user_info_dbobj.resident_city_id})'
+                                              f' does not exist.')
         return city
 
 
@@ -81,8 +80,10 @@ class UserInfo(UserInfoBase):
             return 1
 
     def set_gender(self, gender):
-        if not gender in db_user.UserInfo.GENDER_CHOICES:
-            raise GenderDoNotExistException
+        if gender not in (db_user.UserInfo.MALE,
+                          db_user.UserInfo.FEMALE,
+                          db_user.UserInfo.UNKNOWN):
+            gender = db_user.UserInfo.OTHER
         try:
             self.user_info_dbobj.gender = gender
             self.user_info_dbobj.save()
@@ -91,10 +92,9 @@ class UserInfo(UserInfoBase):
             raise e
             return 1
 
-    def set_resident_city(self, city_id):
+    def set_resident_city_id(self, city_id):
         if not db_user.City.objects.filter(city_id=city_id).exists():
-            raise CityIdDoNotExistException(
-                'City whose ID=%s do not exist.' % city_id)
+            raise CityIdDoesNotExistException(f'City (ID={city_id}) do not exist.')
         try:
             self.user_info_dbobj.resident_city_id = city_id
             self.user_info_dbobj.save()
@@ -107,11 +107,12 @@ class UserInfo(UserInfoBase):
 class FriendInfo(UserInfoBase):
     def __init__(self, user_id, friend_user_id):
         try:
-            self.friend_ralation_dbobj = db_user.FriendRelation.objects.get(
-                user_id=user_id, friend_user_id=friend_user_id)
+            self.friend_ralation_dbobj = db_user.FriendRelation.objects.get(user_id=user_id,
+                                                                            friend_user_id=friend_user_id)
         except ObjectDoesNotExist:
-            raise FriendDoNotExistException(
-                "Friend relation between ID=%d and ID=%d doesn't exist." % (user_id, friend_user_id))
+            raise FriendDoesNotExistException(f'Friend relation between '
+                                              f'user (ID={user_id}) and user (ID={friend_user_id})'
+                                              f' does not exist.')
         super().__init__(friend_user_id)
 
     def get_user_note(self):
